+++ .prettierrc.js
@@ -1,0 +1,3 @@
+const { prettier } = require('./dist/packages/linter')
+
+module.exports = prettier