+++ packages/design-system/src/behavior/Collapsible/Collapsible.stories.tsx
@@ -1,0 +1,7 @@
+import type { ComponentStory } from '@storybook/react';
+import Collapsible from './Collapsible';
+
+export const Template: ComponentStory<typeof Collapsible> = (args) => <Collapsible {...args} />;
+
+export const args = {
+}