+++ packages/design-system/src/behavior/Overlay/Overlay.spec.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import { render } from '@testing-library/react';
+
+import Overlay from './Overlay';
+
+describe('Overlay', () => {
+  it('should render successfully', () => {
+    const { baseElement } = render(<Overlay/>);
+    expect(baseElement).toBeTruthy();
+  });
+});