+++ packages/design-system/src/behavior/Overlay/Overlay.stories.tsx
@@ -1,0 +1,7 @@
+import type { ComponentStory } from '@storybook/react';
+import Overlay from './Overlay';
+
+export const Template: ComponentStory<typeof Overlay> = (args) => <Overlay {...args} />;
+
+export const args = {
+}