+++ packages/design-system/src/components/RTooltip/RTooltip.spec.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import { render } from '@testing-library/react';
+
+import RTooltip from './RTooltip';
+
+describe('RTooltip', () => {
+  it('should render successfully', () => {
+    const { baseElement } = render(<RTooltip/>);
+    expect(baseElement).toBeTruthy();
+  });
+});