+++ packages/design-system/src/components/RTooltip/RTooltip.stories.tsx
@@ -1,0 +1,7 @@
+import type { ComponentStory } from '@storybook/react';
+import RTooltip from './RTooltip';
+
+export const Template: ComponentStory<typeof RTooltip> = (args) => <RTooltip {...args} />;
+
+export const args = {
+}