+++ packages/design-system/src/components/RTooltip/index.ts
@@ -1,0 +1,2 @@
+export { default as RTooltip } from './RTooltip';
+export type { RTooltipProps } from './RTooltip.interface';