+++ packages/design-system/src/components/typography/H3/H3.interface.ts
@@ -1,0 +1,15 @@
+import { BaseHTMLAttributes } from 'react';
+
+export enum Color {
+    Grey = 'grey',
+    White = 'white',
+    Disabled = 'disabled',
+    Ghost = 'ghost',
+    Red = 'red',
+    Yellow = 'yellow',
+    Green = 'green',
+  }
+  
+  export interface H3Props extends BaseHTMLAttributes<HTMLDivElement>{
+    color?: `${Color}`;
+  }