+++ packages/design-system/src/components/typography/H3/H3.stories.tsx
@@ -1,0 +1,8 @@
+import type { ComponentStory } from '@storybook/react';
+import H3 from './H3';
+
+export const Template: ComponentStory<typeof H3> = (args) => <H3 {...args} />;
+
+export const args = {
+    children: 'Heading Three'
+}