+++ packages/design-system/src/components/typography/H3/index.ts
@@ -1,0 +1,2 @@
+export { default as H3 } from './H3';
+export type { H3Props } from './H3.interface';