+++ packages/linter/src/prettier.ts
@@ -1,0 +1,14 @@
+export default {
+  "jsxSingleQuote": true,
+  "printWidth": 120,
+  "tabWidth": 2,
+  "singleQuote": true,
+  "trailingComma": "all",
+  "bracketSpacing": true,
+  "jsxBracketSameLine": true,
+  "parser": "typescript",
+  "semi": true,
+  "proseWrap": "never",
+  "arrowParens": "avoid",
+  "endOfLine": "auto"
+}